rest_api/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aniso8601=7.0.0=pypi_0
asgiref=3.4.1=pypi_0
attrs=21.2.0=pypi_0
autobahn=21.3.1=pypi_0
automat=20.2.0=pypi_0
ca-certificates=2021.10.26=hecd8cb5_2
certifi=2021.10.8=py39hecd8cb5_0
cffi=1.15.0=pypi_0
channels=3.0.4=pypi_0
constantly=15.1.0=pypi_0
cryptography=35.0.0=pypi_0
daphne=3.0.2=pypi_0
django=3.2.9=pypi_0
django-debug-toolbar=3.2.2=pypi_0
django-filter=21.1=pypi_0
djangorestframework=3.12.4=pypi_0
graphene=2.1.9=pypi_0
graphene-django=2.15.0=pypi_0
graphene-django-pagination=0.0.3=pypi_0
graphql-core=2.3.2=pypi_0
graphql-relay=2.0.1=pypi_0
hyperlink=21.0.0=pypi_0
idna=3.3=pypi_0
incremental=21.3.0=pypi_0
libcxx=12.0.0=h2f01273_0
libffi=3.3=hb1e8313_2
ncurses=6.3=hca72f7f_2
openssl=1.1.1l=h9ed2024_0
pillow=8.4.0=pypi_0
pip=21.2.4=py39hecd8cb5_0
promise=2.3=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pyjwt=2.3.0=pypi_0
pymysql=1.0.2=pypi_0
pyopenssl=21.0.0=pypi_0
python=3.9.7=h88f2d9e_1
pytz=2021.3=pypi_0
readline=8.1=h9ed2024_0
rx=1.6.1=pypi_0
service-identity=21.1.0=pypi_0
setuptools=58.0.4=py39hecd8cb5_0
singledispatch=3.7.0=pypi_0
six=1.16.0=pypi_0
sqlite=3.36.0=hce871da_0
sqlparse=0.4.2=pypi_0
text-unidecode=1.3=pypi_0
tk=8.6.11=h7bc2e8c_0
twisted=21.7.0=pypi_0
txaio=21.2.1=pypi_0
typing-extensions=4.0.0=pypi_0
tzdata=2021e=hda174b7_0
wheel=0.37.0=pyhd3eb1b0_1
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zope-interface=5.4.0=pypi_0
